Base navbar functionality added
