+++ client/src/index.css
@@ -29,6 +29,25 @@ html, body, #root, .App {
   text-align: center;
 }
 
+.nav-bar {
+  margin: 0;
+  list-style-type: none;
+  background-color: #F4F1EA;
+  padding: 0;
+}
+
+.nav-bar-content {
+  margin-left: 20px;
+  display: flex;
+  justify-content: flex-start;
+  align-items: center;
+}
+
+.logo {
+  color: brown;
+  font-size: 20px;
+}
+
 form {
   display: flex;
   flex-direction: column;
@@ -158,7 +177,7 @@ button {
   display: block;
 }
 
-a {
+.login a, .signup a {
   margin-left: 10px;
   font-weight: normal;
   color: #00635d;
